SERVEUR_MULTIMEDIA/requirements.txt
#pip install -r requirements.txt
Flask==2.3.2
numpy~=1.26.4
tensorflow~=2.16.1
pillow~=10.3.0